pyproject.toml
[tool.poetry]
name = "plant-disease"
version = "0.1.0"
description = ""
authors = ["Michał Zaręba <michal.zareba@softwaremill.pl>"]
readme = "README.md"
packages = [
    { include = "src", from = "." }
]

[tool.poetry.dependencies]
python = "^3.10"
torch = {version = "2.5.1+cu121", source = "torch"}
ipykernel = "^6.29.5"
numpy = "2.0"
torchvision = {version = "0.20.1+cu121", source = "torch"}
matplotlib = "^3.9.3"
opencv-python = "^4.10.0.84"
helper = "^2.5.0"
tqdm = "^4.67.1"
pre-commit = "^4.0.1"

pyqt6 = "^6.7.1"
pillow = "^11.0.0"
hydra-core = "^1.3.2"
lightning = "^2.5.0.post0"
scikit-learn = "^1.6.0"
mlflow = "^2.19.0"
ultralytics = "^8.3.71"



[[tool.poetry.source]]
name = "PyPI"
priority = "primary"


[[tool.poetry.source]]
name = "torch"
url = "https://download.pytorch.org/whl/cu121"
priority = "supplemental"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
